docs(vip): bilingual STS Game.dll IP patcher usage
tools(vip): add STS client IP patcher

feat(vip): serve China membership stage over STS

fix(path): repath NPC chase when target drifts

Invalidate intermediate chase paths once the player moves far from the
cached path end, while keeping old waypoints so the NPC keeps moving
during async repath.


fix(vip): align China membership benefits

fix(path): stabilize NPC pursuit movement

docs: record retail AI client validation

feat: add independent account VIP support

fix(nav): preserve selected path layer height

data: refine retail NPC roaming spawns

data: sync retail NPC spawns

fix(nav): stabilize A-star NPC pathfinding

fix(nav): extend fixed-location path timeout

fix: apply retail npc drop adjustments

fix(nav): stabilize NPC path height and stopping

fix(combat): apply retail NPC abnormal immunities

feat(retail): migrate AI and quest runtime

fix: restore compact skill execution

fix(path): align retail routing behavior

refactor(schema): move skill schemas

fix(skill): restore item-referenced templates

refactor(data): move skill schema to definitions

feat(skill): migrate retail skill definitions

fix(path): align npc chase and return behavior

feat(path): add retail npc path behavior data

fix(path): improve spatial routing and crowd avoidance

fix(path): preserve terrain and crowd timing

fix(path): guard asynchronous route state

fix(path): make npc routing asynchronous

feat(geo): load compressed path and world data

test(path): verify queue and crowd behavior

data(path): add retail path maps

feat(path): add spatial routing and crowd coordination

fix(trade): restore vendor buy/sell price formulas

Hardcoded *100 broke PricesService and TradeList, inflating shop prices.
Restore global price/modifier usage and count*modifier/100 for buy totals.


feat(geo): replace navmesh with retail pathfinding

config(drop): disable global drops by default

全局掉落当前只额外掉 Kinah，默认关掉，避免怪物额外出金币。

refactor(data): split world fly path definitions

refactor(data): move hotspot locations to compact definitions

Load hotspot_location from definitions/compact and keep the existing
JAXB template shape, matching the npc_drops compact layout.

refactor(data): split windpaths into compact wind.xml

Move windpath definitions out of world.xml and load them from the
new compact wind.xml alongside fly_path metadata.

fix(data): sync hotspot locations from zonemap_hotspot

Align hotspot coords, heading, and base price with compact retail
zonemap data while keeping legacy-only entries and the existing XML
schema.

refactor(data): move npc drops to compact definitions

fix(combat): correct reflected damage handling

fix(data): validate static data entry point

fix: align godstone activation behavior

revert: use compact npc skill definitions

revert: restore legacy npc skill data

docs: add retail skill migration plan

+++ src/main/java/com/aionemu/gameserver/utils/MathUtil.java
@@ -392,7 +392,7 @@ public class MathUtil {
 	 *
 	 * @return 若 near 则为 true / True if near
 	 */
-	public final static boolean isNearCoordinates(VisibleObject obj, float x, float y, float z, int offset) {
+	public final static boolean isNearCoordinates(VisibleObject obj, float x, float y, float z, float offset) {
 		return getDistance(obj.getX(), obj.getY(), obj.getZ(), x, y, z) < offset + NpcMoveController.MOVE_CHECK_OFFSET;
 	}
